copy all configurations from csbc-pson to cancercomplexity portal, and fix scripts to point to the new bucket
