+++ src/configurations/cancercomplexity/betaBannerConfig.ts
@@ -1,5 +1,4 @@
 import { BetaBanner } from 'types/portal-config'
 
-export const betaBanner: BetaBanner = {
-  // TODO: Fill in with properties if a beta banner is desired
-}
+const betaBanner: BetaBanner = {}
+export default betaBanner
