+++ src/configurations/cancercomplexity/docTitleConfig.ts
@@ -1,8 +1,9 @@
 import { DocTitleConfig } from 'types/portal-config'
 
 const docTitleConfig: DocTitleConfig = {
-  name: 'TODO',
-  description: 'TODO',
+  name: 'Cancer Complexity Knowledge Portal',
+  description:
+    'The NIH National Cancer Institute-sponsored Cancer Systems Biology Consortium (CSBC) and Physical Sciences in Oncology Network (PS-ON) aim to tackle the most perplexing issues in cancer to increase our understanding of tumor biology, treatment options, and patient outcome. The CSBC / PS-ON Data Portal highlights ongoing research efforts in these richly interdisciplinary fields, including the data generated and methods developed in the addresssing fundamental questions in cancer research.',
 }
 
 export default docTitleConfig
