+++ src/configurations/cancercomplexity/footerConfig.ts
@@ -1,8 +1,9 @@
 import { FooterConfig } from 'types/portal-config'
 
 const footer: FooterConfig = {
-  termsOfService: 'TODO',
-  contactUs: 'TODO'
+  termsOfService:
+    'https://s3.amazonaws.com/static.synapse.org/governance/SageBionetworksSynapseTermsandConditionsofUse.pdf?v=5',
+  contactUs: 'mailto:csbc_pson_dcc@sagebase.org',
 }
 
 export default footer
