--- src/configurations/cancercomplexity/headerConfig.ts
@@ -1,8 +1,0 @@
-import { HomePageHeaderConfig } from 'types/portal-config'
-
-const homePageHeader: HomePageHeaderConfig = {
-  summary: 'TODO',
-  title: 'TODO'
-}
-
-export default homePageHeader
