+++ src/configurations/cancercomplexity/loadingScreen.tsx
@@ -1,6 +1,10 @@
 import * as React from 'react'
 import { BarLoader } from 'react-spinners'
 
-const loadingScreen = <div className="bar-loader"><BarLoader color="#47337D" loading={true}/></div>
+const loadingScreen = (
+  <div className="bar-loader">
+    <BarLoader color="#47337D" loading={true} />
+  </div>
+)
 
 export default loadingScreen
