+++ src/configurations/cancercomplexity/logoFooterConfig.ts
@@ -1,6 +1,5 @@
 import { LogoConfig } from 'types/portal-config'
-
-const logoHeader: LogoConfig = {
-  name: 'TODO'
+const logo: LogoConfig = {
+  name: 'Cancer Complexity Knowledge Portal',
 }
-export default logoHeader
+export default logo
