+++ src/configurations/cancercomplexity/style/_overides.scss
@@ -1,18 +1,35 @@
+$primary-text-color: #407ba0;
+$primary-action-color: #47337d;
 
-// $primary-text-color: ;// TODO
-// $primary-action-color: ; // TODO
+#header,
+#footer {
+  background-blend-mode: multiply;
+}
 
-// // Note length($button-colors) must equal length of explore page routes
-// $button-colors: (
-// );
+.SRC-chart-link {
+  a {
+    background: $primary-action-color !important;
+  }
+}
 
-// // Since scss breaks if a file is missing in a local url link these functinons must get manually overwitten
-// @function getHeader($path) {
-//   // TODO
-//   @return url($path + "./header.svg");
-// }
+.CSBC-ThemesPlot {
+  h2.title {
+    margin-top: 63px;
+    text-align: center;
+    text-transform: none;
+    margin-bottom: 30px;
+    font-size: 24px;
+    font-weight: bold;
+    color: $primary-text-color !important;
+  }
+  margin-bottom: 63px;
+}
 
-// @function getFooter($path) {
-//   // TODO
-//   @return "";
-// }
+// Since scss breaks if a file is missing in a local url link these functinons must get manually overwitten
+@function getHeader($path) {
+  @return url($path + '/dna_header.svg');
+}
+
+@function getFooter($path) {
+  @return url($path + '/dna.svg');
+}
