+++ src/configurations/cancercomplexity/synapseConfigs/commonProps.ts
