+++ src/configurations/cancercomplexity/synapseConfigs/index.ts
@@ -1,4 +1,10 @@
-// TODO: Export configurations
-export { }
+import { files } from './files'
+import { datasets } from './datasets'
+import { grants } from './grants'
+import { publications } from './publications'
+import { projects } from './projects'
+import { tools } from './tools'
 
-export default { }
+export { files, datasets, grants, publications, projects, tools }
+
+export default { files, datasets, grants, publications, projects, tools }
