+++ src/configurations/cancercomplexity/synapseConfigs/onPointClick.ts
