--- src/configurations/csbc-pson/betaBannerConfig.ts
@@ -1,4 +1,0 @@
-import { BetaBanner } from 'types/portal-config'
-
-const betaBanner: BetaBanner = {}
-export default betaBanner
