--- src/configurations/csbc-pson/footerConfig.ts
@@ -1,9 +1,0 @@
-import { FooterConfig } from 'types/portal-config'
-
-const footer: FooterConfig = {
-  termsOfService:
-    'https://s3.amazonaws.com/static.synapse.org/governance/SageBionetworksSynapseTermsandConditionsofUse.pdf?v=5',
-  contactUs: 'mailto:csbc_pson_dcc@sagebase.org',
-}
-
-export default footer
