--- src/configurations/csbc-pson/loadingScreen.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react'
-import { BarLoader } from 'react-spinners'
-
-const loadingScreen = (
-  <div className="bar-loader">
-    <BarLoader color="#47337D" loading={true} />
-  </div>
-)
-
-export default loadingScreen
