--- src/configurations/csbc-pson/logoFooterConfig.ts
@@ -1,5 +1,0 @@
-import { LogoConfig } from 'types/portal-config'
-const logo: LogoConfig = {
-  name: 'Cancer Complexity Knowledge Portal',
-}
-export default logo
