--- src/configurations/csbc-pson/logoHeaderConfig.ts
@@ -1,6 +1,0 @@
-import { LogoConfig } from 'types/portal-config'
-
-const logoHeader: LogoConfig = {
-  name: 'Cancer Complexity Knowledge Portal',
-}
-export default logoHeader
