--- src/configurations/csbc-pson/scripts/exportS3ProductionBucketName.sh
@@ -1,1 +1,0 @@
-export S3_PRODUCTION_BUCK_LOCATION=s3://prod.csbc-pson.synapse.org
