--- src/configurations/csbc-pson/scripts/exportS3StagingBucketName.sh
@@ -1,1 +1,0 @@
-export S3_STAGING_BUCKET_LOCATION=s3://staging.csbc-pson.synapse.org
