--- src/configurations/csbc-pson/style/_overides.scss
@@ -1,35 +1,0 @@
-$primary-text-color: #407ba0;
-$primary-action-color: #47337d;
-
-#header,
-#footer {
-  background-blend-mode: multiply;
-}
-
-.SRC-chart-link {
-  a {
-    background: $primary-action-color !important;
-  }
-}
-
-.CSBC-ThemesPlot {
-  h2.title {
-    margin-top: 63px;
-    text-align: center;
-    text-transform: none;
-    margin-bottom: 30px;
-    font-size: 24px;
-    font-weight: bold;
-    color: $primary-text-color !important;
-  }
-  margin-bottom: 63px;
-}
-
-// Since scss breaks if a file is missing in a local url link these functinons must get manually overwitten
-@function getHeader($path) {
-  @return url($path + '/dna_header.svg');
-}
-
-@function getFooter($path) {
-  @return url($path + '/dna.svg');
-}
