--- src/configurations/csbc-pson/synapseConfigs/index.ts
@@ -1,10 +1,0 @@
-import { files } from './files'
-import { datasets } from './datasets'
-import { grants } from './grants'
-import { publications } from './publications'
-import { projects } from './projects'
-import { tools } from './tools'
-
-export { files, datasets, grants, publications, projects, tools }
-
-export default { files, datasets, grants, publications, projects, tools }
